Set up k8s API server only when first needed for a test
As part of this, the k8s API server setup/teardown code is also moved
into a subpackage.
